% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pipeline.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Create a lasR pipeline to do all the normal stuff}
\usage{
create_pipeline(output_dir)
}
\arguments{
\item{output_dir}{directory to save the results}
}
\value{
a lasR pipeline
}
\description{
Create a lasR pipeline to do all the normal stuff
}
\examples{
pipeline = create_pipeline(".")

}
