% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pipeline.R
\name{setup_pipeline}
\alias{setup_pipeline}
\title{Setup directories for pipeline}
\usage{
setup_pipeline(output_dir)
}
\arguments{
\item{output_dir}{directory to save the results}
}
\value{
path to output_dir
}
\description{
Setup directories for pipeline
}
